+++ crates/blockifier/src/blockifier/stateful_validator.rs
@@ -19,8 +19,7 @@ use crate::state::errors::StateError;
 use crate::state::state_api::StateReader;
 use crate::transaction::account_transaction::AccountTransaction;
 use crate::transaction::errors::{TransactionExecutionError, TransactionPreValidationError};
-use crate::transaction::transaction_execution::Transaction;
-use crate::transaction::transactions::ValidatableTransaction;
+use crate::transaction::transactions::{Transaction, ValidatableTransaction};
 
 #[cfg(test)]
 #[path = "stateful_validator_test.rs"]
