+++ crates/blockifier/src/blockifier/transaction_executor.rs
@@ -22,8 +22,7 @@ use crate::state::errors::StateError;
 use crate::state::state_api::StateReader;
 use crate::transaction::errors::TransactionExecutionError;
 use crate::transaction::objects::TransactionExecutionInfo;
-use crate::transaction::transaction_execution::Transaction;
-use crate::transaction::transactions::{ExecutableTransaction, ExecutionFlags};
+use crate::transaction::transactions::{ExecutableTransaction, ExecutionFlags, Transaction};
 
 #[cfg(test)]
 #[path = "transaction_executor_test.rs"]
