+++ crates/blockifier/src/concurrency/worker_logic.rs
@@ -23,8 +23,7 @@ use crate::state::cached_state::{
 };
 use crate::state::state_api::{StateReader, UpdatableState};
 use crate::transaction::objects::{TransactionExecutionInfo, TransactionExecutionResult};
-use crate::transaction::transaction_execution::Transaction;
-use crate::transaction::transactions::{ExecutableTransaction, ExecutionFlags};
+use crate::transaction::transactions::{ExecutableTransaction, ExecutionFlags, Transaction};
 
 #[cfg(test)]
 #[path = "worker_logic_test.rs"]
